refactor: move game struct and its data to external file
