--- SteamList/Game.swift
@@ -1,8 +1,0 @@
-//
-//  Game.swift
-//  SteamList
-//
-//  Created by Macbook on 30/12/2022.
-//
-
-import Foundation
