+++ SteamList/MockData.swift
@@ -1,0 +1,41 @@
+//
+//  MockData.swift
+//  SteamList
+//
+//  Created by Macbook on 30/12/2022.
+//
+
+import Foundation
+
+class MockData {
+    static var games: [Game] = [
+    Game(title: "Dota 2", isFavorite: false),
+    Game(title: "League of Legends", isFavorite: false),
+    Game(title: "GTA Vice City", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Counter Strkie 1.6", isFavorite: false),
+    Game(title: "Need For Speed", isFavorite: false)
+    ]
+}
